+++ plugins/fr.inria.diverse.puzzle.metrics/src/fr/inria/diverse/puzzle/metrics/evaluators/syntax/TotalAmountOfMethods.java
@@ -11,7 +11,7 @@ import fr.inria.diverse.melange.metamodel.melange.Language;
 
 public class TotalAmountOfMethods {
 
-	public static double evaluateMetric(ArrayList<Language> languages, ConceptComparison conceptComparisonOperator, MethodComparison methodComparisonOperator ){
+	public static int evaluateMetric(ArrayList<Language> languages, ConceptComparison conceptComparisonOperator, MethodComparison methodComparisonOperator ){
 		ArrayList<ConceptMethodMemberVO> conceptMethodMemberList = FamiliesServices.getInstance().getConceptMethodMemberMappingList(languages);
 		ArrayList<ConceptMethodMembersGroupVO> conceptMethodMemberGroupList = FamiliesServices.getInstance().getConceptMethodMemberGroupList(conceptMethodMemberList, conceptComparisonOperator, methodComparisonOperator);
 		return conceptMethodMemberGroupList.size();
